test(lint): clean up static analysis warnings in tests
