+++ tests/test_chart_builder.py
@@ -8,7 +8,6 @@ import warnings
 from typing import cast
 
 import pandas as pd
-import plotly.graph_objects as go
 
 from analyze_git_repo_loc.reporting.chart_builder import ChartBuilder, ChartStrategy
 
@@ -44,10 +43,10 @@ def test_weekly_trend_chart_uses_span_based_ticks_for_sparse_long_ranges() -> No
         )
     )
 
-    xaxis = cast(go.layout.XAxis, fig.layout.xaxis)
+    xaxis = cast(dict[str, object], fig.layout.to_plotly_json()["xaxis"])
 
-    assert xaxis.dtick == "M1"
-    assert xaxis.tickformat == "%b %Y"
+    assert xaxis["dtick"] == "M1"
+    assert xaxis["tickformat"] == "%b %Y"
 
 
 def test_daily_trend_chart_uses_sparse_ticks_for_medium_ranges() -> None:
@@ -65,14 +64,15 @@ def test_daily_trend_chart_uses_sparse_ticks_for_medium_ranges() -> None:
         )
     )
 
-    xaxis = cast(go.layout.XAxis, fig.layout.xaxis)
+    xaxis = cast(dict[str, object], fig.layout.to_plotly_json()["xaxis"])
+    tickvals = cast(list[object], xaxis["tickvals"])
 
-    assert xaxis.dtick == "D14"
-    assert xaxis.tickformat == "%b %d"
-    assert xaxis.tickmode == "array"
-    assert len(xaxis.tickvals) <= 10
-    assert xaxis.tickvals[0] == dates[0]
-    assert xaxis.tickvals[-1] == dates[-1]
+    assert xaxis["dtick"] == "D14"
+    assert xaxis["tickformat"] == "%b %d"
+    assert xaxis["tickmode"] == "array"
+    assert len(tickvals) <= 10
+    assert tickvals[0] == dates[0]
+    assert tickvals[-1] == dates[-1]
 
 
 def test_trend_chart_build_does_not_emit_plotly_deprecation_warning() -> None:
